Extract functionality of Statement page into custom hook and Statementform component
